Add files via upload
Almost nothing changed, apart from some normalization, that is not assured to be correct (e.g. always forcing the bias to be positive is... rough.)

Still to solve: when should one center and rescale the data? Before each SVM? Otherwise the results can become quite wild... (and many neurons dead).